Major review of Abstractions and CodeAreas (and related).
* The arity is stored in the CodeArea instead of the Abstraction.
  (It is however cached in Abstraction like other data.)

* Arities and Xcounts are typed size_t instead of int.

* CodeAreas store a print name and additional debug data. The
  print name is used by Abstraction::printReprToStream().

+++ vm/vm/main/builtins.hh
@@ -44,7 +44,7 @@ UnstableNode BaseBuiltin::getNameAtom(VM vm) {
 }
 
 OpResult BaseBuiltin::getCallInfo(
-  RichNode self, VM vm, int& arity, ProgramCounter& start, int& Xcount,
+  RichNode self, VM vm, size_t& arity, ProgramCounter& start, size_t& Xcount,
   StaticArray<StableNode>& Gs, StaticArray<StableNode>& Ks) {
 
   if (_codeBlock == nullptr)
